Ms2ex.Context.Emotes (MS2EX v0.1.0)

Context module for character emote-related operations.

This module provides functions for listing, learning, and accessing default emotes available to characters.

Functions

default_emotes()

@spec default_emotes() :: [integer()]

Returns the list of default emote IDs available to all characters.

Examples

iex> default_emotes()
[90200011, 90200004, 90200024, 90200041, 90200042]

learn(character, emote_id)

@spec learn(Ms2ex.Schema.Character.t(), integer()) ::
  {:ok, Ms2ex.Schema.Emote.t()} | {:error, Ecto.Changeset.t()}

Makes a character learn a new emote by ID.

Examples

iex> learn(character, 90200011)
{:ok, %Schema.Emote{}}

iex> learn(character, 90200011) # when already learned
{:error, %Ecto.Changeset{}}

list(character)

@spec list(Ms2ex.Schema.Character.t()) :: [integer()]

Lists all emote IDs that a character has learned.

Examples

iex> list(character)
[90200011, 90200004, 90200024, 90200041, 90200042]